requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
python-dotenv==1.0.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0



